requirements.txt
attrs==21.2.0
backports.entry-points-selectable==1.1.1
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
colorama==0.4.4
cycler==0.11.0
distlib==0.3.3
et-xmlfile==1.1.0
filelock==3.4.0
Flask==2.0.2
Flask-Cors==3.0.10
flask-marshmallow==0.14.0
Flask-MySQLdb==0.2.0
Flask-SQLAlchemy==2.5.1
fonttools==4.28.5
google-auth==2.3.3
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.10.1
itsdangerous==2.0.1
Jinja2==3.0.3
joblib==1.1.0
jsonschema==4.3.2
kiwisolver==1.3.2
Markdown==3.3.6
MarkupSafe==2.0.1
marshmallow==3.14.1
marshmallow-sqlalchemy==0.27.0
matplotlib==3.5.1
mysqlclient==2.1.0
numpy==1.22.1
oauthlib==3.1.1
openpyxl==3.0.9
packaging==21.3
pandas==1.3.5
Pillow==9.0.0
platformdirs==2.4.0
protobuf==3.19.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pygame==2.1.0
Pygments==2.10.0
PyMySQL==1.0.2
pyparsing==3.0.6
pyrsistent==0.18.0
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.8
scikit-learn==1.0.2
scipy==1.7.3
seaborn==0.11.2
six==1.16.0
sklearn==0.0
SQLAlchemy==1.4.29
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
threadpoolctl==3.0.0
typing_extensions==4.0.1
urllib3==1.26.7
virtualenv==20.10.0
Werkzeug==2.0.2
xlrd==2.0.1
zipp==3.7.0
